+++ crates/rust-client/src/accounts.rs
@@ -15,11 +15,14 @@ use miden_objects::{
     accounts::AuthSecretKey,
     assets::TokenSymbol,
     crypto::{dsa::rpo_falcon512::SecretKey, rand::FeltRng},
-    Felt, Word,
+    Digest, Felt, Word,
 };
 
 use super::Client;
-use crate::ClientError;
+use crate::{
+    store::{AccountRecord, AccountStatus},
+    ClientError,
+};
 
 /// Defines templates for creating different types of Miden accounts.
 pub enum AccountTemplate {
@@ -190,36 +193,27 @@ impl<R: FeltRng> Client<R> {
     // ACCOUNT DATA RETRIEVAL
     // --------------------------------------------------------------------------------------------
 
-    /// Returns a list of [AccountHeader] of all accounts stored in the database along with the
-    /// seeds used to create them.
+    /// Returns a list of [AccountHeader] of all accounts stored in the database along with their
+    /// statuses.
     ///
     /// Said accounts' state is the state after the last performed sync.
     pub async fn get_account_headers(
         &self,
-    ) -> Result<Vec<(AccountHeader, Option<Word>)>, ClientError> {
+    ) -> Result<Vec<(AccountHeader, AccountStatus)>, ClientError> {
         self.store.get_account_headers().await.map_err(|err| err.into())
     }
 
-    /// Retrieves a full [Account] object. The seed will be returned if the account is new,
-    /// otherwise it will be `None`.
-    ///
-    /// This function returns the [Account]'s latest state. If the account is new (that is, has
-    /// never executed a transaction), the returned seed will be `Some(Word)`; otherwise the seed
-    /// will be `None`
+    /// Retrieves a full [AccountRecord] object for the specified `account_id`. This result
+    /// represents data for the latest state known to the client, alongside its status.
     ///
     /// # Errors
     ///
     /// Returns a `StoreError::AccountDataNotFound` if there is no account for the provided ID
-    pub async fn get_account(
-        &self,
-        account_id: AccountId,
-    ) -> Result<(Account, Option<Word>), ClientError> {
+    pub async fn get_account(&self, account_id: AccountId) -> Result<AccountRecord, ClientError> {
         self.store.get_account(account_id).await.map_err(|err| err.into())
     }
 
-    /// Retrieves an [AccountHeader] object for the specified [AccountId] along with the seed
-    /// used to create it. The seed will be returned if the account is new, otherwise it
-    /// will be `None`.
+    /// Retrieves an [AccountHeader] object for the specified [AccountId] along with its status.
     ///
     /// Said account's state is the state according to the last sync performed.
     ///
@@ -229,7 +223,7 @@ impl<R: FeltRng> Client<R> {
     pub async fn get_account_header_by_id(
         &self,
         account_id: AccountId,
-    ) -> Result<(AccountHeader, Option<Word>), ClientError> {
+    ) -> Result<(AccountHeader, AccountStatus), ClientError> {
         self.store.get_account_header(account_id).await.map_err(|err| err.into())
     }
 
@@ -245,6 +239,40 @@ impl<R: FeltRng> Client<R> {
         account_id: AccountId,
     ) -> Result<AuthSecretKey, ClientError> {
         self.store.get_account_auth(account_id).await.map_err(|err| err.into())
+    }
+}
+
+// ACCOUNT UPDATES
+// ================================================================================================
+
+/// Contains account changes to apply to the store.
+pub struct AccountUpdates {
+    /// Updated public accounts.
+    updated_onchain_accounts: Vec<Account>,
+    /// Node account hashes that do not match the tracked information.
+    mismatched_offchain_accounts: Vec<(AccountId, Digest)>,
+}
+
+impl AccountUpdates {
+    /// Creates a new instance of `AccountUpdates`.
+    pub fn new(
+        updated_onchain_accounts: Vec<Account>,
+        mismatched_offchain_accounts: Vec<(AccountId, Digest)>,
+    ) -> Self {
+        Self {
+            updated_onchain_accounts,
+            mismatched_offchain_accounts,
+        }
+    }
+
+    /// Returns the updated public accounts.
+    pub fn updated_onchain_accounts(&self) -> &[Account] {
+        &self.updated_onchain_accounts
+    }
+
+    /// Returns the mismatched offchain accounts.
+    pub fn mismatched_offchain_accounts(&self) -> &[(AccountId, Digest)] {
+        &self.mismatched_offchain_accounts
     }
 }