+++ crates/rust-client/src/transactions/mod.rs
@@ -33,7 +33,7 @@ use crate::{
     notes::{NoteScreener, NoteUpdates},
     store::{
         input_note_states::ExpectedNoteState, InputNoteRecord, InputNoteState, NoteFilter,
-        OutputNoteRecord, TransactionFilter,
+        OutputNoteRecord, StoreError, TransactionFilter,
     },
     sync::NoteTagRecord,
     ClientError,
@@ -459,9 +459,20 @@ impl<R: FeltRng> Client<R> {
 
         let account_id = tx_result.executed_transaction().account_id();
         let account_delta = tx_result.account_delta();
-        let (mut account, _seed) = self.get_account(account_id).await?;
+        let account_record = self.get_account(account_id).await?;
 
+        if account_record.is_locked() {
+            return Err(ClientError::AccountLocked(account_id));
+        }
+
+        let mut account: Account = account_record.into();
         account.apply_delta(account_delta)?;
+
+        if self.store.get_account_header_by_hash(account.hash()).await?.is_some() {
+            return Err(ClientError::StoreError(StoreError::AccountHashAlreadyExists(
+                account.hash(),
+            )));
+        }
 
         // Save only input notes that we care for (based on the note screener assessment)
         let created_input_notes = tx_result.relevant_notes().to_vec();
@@ -658,7 +669,8 @@ impl<R: FeltRng> Client<R> {
         account_id: AccountId,
         transaction_request: &TransactionRequest,
     ) -> Result<(), ClientError> {
-        let (account, _) = self.get_account(account_id).await?;
+        let account: Account = self.get_account(account_id).await?.into();
+
         if account.is_faucet() {
             // TODO(SantiagoPittella): Add faucet validations.
             Ok(())
@@ -672,7 +684,7 @@ impl<R: FeltRng> Client<R> {
         &self,
         account_id: AccountId,
     ) -> Result<AccountCapabilities, ClientError> {
-        let account = self.get_account(account_id).await?.0;
+        let account: Account = self.get_account(account_id).await?.into();
         let account_auth = self.get_account_auth(account_id).await?;
 
         // TODO: we should check if the account actually exposes the interfaces we're trying to use