+++ tests/integration/common.rs
@@ -288,7 +288,7 @@ pub async fn assert_account_has_single_asset(
     asset_account_id: AccountId,
     expected_amount: u64,
 ) {
-    let (regular_account, _seed) = client.get_account(account_id).await.unwrap();
+    let regular_account: Account = client.get_account(account_id).await.unwrap().into();
 
     assert_eq!(regular_account.vault().assets().count(), 1);
     let asset = regular_account.vault().assets().next().unwrap();