+++ tests/integration/onchain_tests.rs
@@ -138,7 +138,8 @@ async fn test_onchain_accounts() {
     let second_client_target_account_id = second_client_first_regular_account.id();
     let faucet_account_id = faucet_account_header.id();
 
-    let (_, faucet_seed) = client_1.get_account_header_by_id(faucet_account_id).await.unwrap();
+    let (_, status) = client_1.get_account_header_by_id(faucet_account_id).await.unwrap();
+    let faucet_seed = status.seed().cloned();
     let auth_info = client_1.get_account_auth(faucet_account_id).await.unwrap();
     client_2
         .insert_account(&faucet_account_header, faucet_seed, &auth_info)
@@ -204,7 +205,7 @@ async fn test_onchain_accounts() {
         .get_account(from_account_id)
         .await
         .unwrap()
-        .0
+        .account()
         .vault()
         .get_balance(faucet_account_id)
         .unwrap_or(0);
@@ -212,7 +213,7 @@ async fn test_onchain_accounts() {
         .get_account(to_account_id)
         .await
         .unwrap()
-        .0
+        .account()
         .vault()
         .get_balance(faucet_account_id)
         .unwrap_or(0);
@@ -254,7 +255,7 @@ async fn test_onchain_accounts() {
         .get_account(from_account_id)
         .await
         .unwrap()
-        .0
+        .account()
         .vault()
         .get_balance(faucet_account_id)
         .unwrap_or(0);
@@ -262,7 +263,7 @@ async fn test_onchain_accounts() {
         .get_account(to_account_id)
         .await
         .unwrap()
-        .0
+        .account()
         .vault()
         .get_balance(faucet_account_id)
         .unwrap_or(0);