+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* [BREAKING] Added `AccountRecord` with information about the account's status (#600).
 * Added caching for foreign account code (#597).
 
 ### Fixes